Template pages
a few weeks to the way the template pages work so that we can copy
default pages to groups when they are created. New groups should now
get a default forum, schedule, blog and if set up you can set up two
site pages and tell it to add them for new groups, one a s a template
one as a default course page. needed group blogs to work for this first.