json newsfeed loading
Loads newsfeed via son and then when you scroll to the bottom it loads
some more. still need to figure out how to navigate away and back
without fucking where we are i.e. once i’ve scrolled down a ways and
then click on something I need to go back to what I clicked on not back
to the start.