Update report to use new assessment & Add visibility check
Visibility checkbox allows you to set an assessment to non visibility
which means no one sees the grade!!! it does show up in the report but
thats it. Update to use the new assessment artefact instead of the old
clunky block type, although it will fall back to that if necessary for
older pages.