Bulk user added archive files
There is a problem where loading the files in the file browser causes
things to slow down too much and uploads stopped working. So I have
written a function to move all files without a folder into an archive
folder and also removed the necessary count(children) from the sql
which was taking forever. this speeds things up and sorts out uploading
problems